Get rid of "unknown piece nil" errors when drawing shapes
The logic for drawing the bottom layer was flawed:

It used to be a problem to draw the layer correctly due to the order of the
pieces in that layer. To draw the pieces in the correct order, the layer had to
be split at 6, and then the right hand side would have to be drawn first.

This logic was greatly flawed as all shapes cannot always be sliced at 6.
The logic was reverted to the simpler design there exists now, but the
implementation still tried to split the layer.