Fix slow hotload / compilation when editing draw logic
The code for drawing was originally in the same file as the quil sketch
definition and reagent component definition. I think this made changes to the
file veeeeery slow (molasses mode).

Now the drawing logic is entirely in its own file and any changes to the code
are lightning fast like they should be!